Move results code to executions. WIP Integrate exec reporting
