+++ cli/medperf/entities/execution.py
@@ -18,7 +18,7 @@ class Execution(Entity, ApprovableSchema):
     benchmark: int
     model: int
     dataset: int
-    results: dict
+    results: dict = {}
     metadata: dict = {}
     user_metadata: dict = {}
     model_report: dict = {}
@@ -65,13 +65,18 @@ class Execution(Entity, ApprovableSchema):
         comms_fn = config.comms.get_executions
         if "owner" in filters and filters["owner"] == get_medperf_user_data()["id"]:
             comms_fn = config.comms.get_user_executions
-        if "benchmark" in filters and filters["benchmark"] is not None:
+        elif "benchmark" in filters and filters["benchmark"] is not None:
+            # Only override the communications method if its not related to
+            # the current user
+            # This is needed so that users can filter their own executions
+            # without getting permission errors.
+            # Users will still be able to filter by benchmark through query params
             bmk = filters["benchmark"]
 
-            def get_benchmark_executions():
+            def get_benchmark_executions(*args, **kwargs):
                 # Decorate the benchmark results remote function so it has the same signature
                 # as all the comms_fns
-                return config.comms.get_benchmark_executions(bmk)
+                return config.comms.get_benchmark_executions(bmk, *args, **kwargs)
 
             comms_fn = get_benchmark_executions
